+++ runtime-tests/tests/org_registration.rs
@@ -23,7 +23,7 @@ use radicle_registry_test_utils::*;
 
 #[async_std::test]
 async fn register_org() {
-    let client = Client::new_emulator();
+    let (client, _) = Client::new_emulator();
     let (author, _) = key_pair_with_associated_user(&client).await;
 
     let initial_balance = client.free_balance(&author.public()).await.unwrap();
@@ -60,7 +60,7 @@ async fn register_org() {
 
 #[async_std::test]
 async fn register_with_duplicated_org_id() {
-    let client = Client::new_emulator();
+    let (client, _) = Client::new_emulator();
     let (author, _) = key_pair_with_associated_user(&client).await;
     let register_org_message = random_register_org_message();
 
@@ -76,7 +76,7 @@ async fn register_with_duplicated_org_id() {
 
 #[async_std::test]
 async fn unregister_org() {
-    let client = Client::new_emulator();
+    let (client, _) = Client::new_emulator();
     let (author, _) = key_pair_with_associated_user(&client).await;
 
     let register_org_message = random_register_org_message();
@@ -125,7 +125,7 @@ async fn unregister_org() {
 
 #[async_std::test]
 async fn unregister_org_bad_actor() {
-    let client = Client::new_emulator();
+    let (client, _) = Client::new_emulator();
     let (author, _) = key_pair_with_associated_user(&client).await;
     let register_org_message = random_register_org_message();
 
@@ -174,7 +174,7 @@ async fn unregister_org_bad_actor() {
 
 #[async_std::test]
 async fn unregister_org_with_projects() {
-    let client = Client::new_emulator();
+    let (client, _) = Client::new_emulator();
     let (author, _) = key_pair_with_associated_user(&client).await;
 
     let org_id = random_id();